requirements.txt
# PAM-TALK Project Requirements

# Algorand Integration
py-algorand-sdk>=2.11.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Time Series Forecasting
prophet>=1.1.0

# Machine Learning
scikit-learn>=1.3.0

# HTTP Client
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-html>=3.0.0
pytest-cov>=4.0.0

# Additional utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Phase 1 Performance Optimization
gunicorn>=21.2.0
gevent>=23.7.0
redis>=5.0.0
celery>=5.3.0
flower>=2.0.1

# Performance Testing
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# Monitoring
psutil>=5.9.0